// src/api/api.generated.ts
export enum DTOInvoiceFieldOrder {
  DTOId = "id",
  DTOAmount = "amount",
  DTOStatus = "status",
  DTOLifeTime = "life_time",
  DTODescription = "description",
  DTOPayToAddress = "pay_to_address",
  DTOPaidByAddress = "paid_by_address",
  DTODateCreate = "date_create",
  DTODatePaid = "date_paid",
}

export enum DTOCryptoCurrency {
  DTO_TON = "TON",
  DTO_USDT = "USDT",
}

/** @default "mainnet" */
export enum DTOChain {
  DTOMainnet = "mainnet",
  DTOTestnet = "testnet",
}

/** @example "pending" */
export enum DTOInvoiceStatus {
  DTOPending = "pending",
  DTOPaid = "paid",
  DTOCancelled = "cancelled",
  DTOExpired = "expired",
}
// ...
export interface DTOInvoicesInvoice {
  /** @example "60ffb075" */
  id: string;
  /** @example "1000000000" */
  amount: string;
  /** @example "1000000000" */
  overpayment?: string;
  /** @example "Test description" */
  description: string;
  status: DTOInvoiceStatus;
  /** @example "0:97146a46acc2654y27947f14c4a4b14273e954f78bc017790b41208b0043200b" */
  pay_to_address: string;
  /** @example "0:97146a46acc2654y27947f14c4a4b14273e954f78bc017790b41208b0043200b" */
  paid_by_address?: string;
  /** @example "https://app.tonkeeper.com/transfer/UQ....UQ" */
  payment_link: string;
  /** @example {"first_key":"1","second_key":2} */
  info?: any;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_change: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_expire: number;
  /**
   * @format int64
   * @example 1690889913000
   */
  date_create: number;
  currency: DTOCryptoCurrency;
}
// ...
export enum DTOGetInvoicesParamsTypeOrderEnum {
  DTOAsc = "asc",
  DTODesc = "desc",
}